feat: Enable not repeating countingSystemId
countingSystemId exists both in the payload and the MQTT topic. That was
done so that the MQTT topic structure could later be changed to not
include countingSystemId. Telia wished to avoid creating two versions of
their software, one for HSL and one for Waltti, so we agreed to read the
countingSystemId from the topic if it is not available in the payload.